requirements.txt
starlite[cryptography]==1.38.0
python-jose==3.3.0
uvicorn==0.19.0
SQLAlchemy==1.4.43
Pillow==9.3.0
segno==1.5.2
qrcode-artistic==2.1.0
requests==2.28.1
UliEngineering==0.3.10
pydantic==1.10.2